+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [0.0.2] - 2020/10/30
+* Add get screen height, screen width, status bar height.
+1. ResponsiveUnit.screenWidth 
+2. ResponsiveUnit.screenHeight
+3. ResponsiveUnit.statusBarHeight
+
 ## [0.0.1] - 2020/09/13
 * Init project, add flutter responsive layout unit
 1. rpt